requirements.txt
keras
matplotlib
numpy
pandas
pandas_datareader
Pillow
scikit_learn
datetime
streamlit
yfinance